pyproject.toml
[project]
name = "coder_ai"
version = "0.1.0"
description = "Coder_AI using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.105.0,<1.0.0",
    "crawl4ai>=0.1.0,<1.0.0",
    "python-dotenv>=1.0.0",
    "docling>=1.0.0",
    
]

[project.scripts]
kickoff = "coder_ai.main:kickoff"
plot = "coder_ai.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
